@@ -2,11 +2,12 @@
 composite construct that sets up all resources
 for SSO group life cycle notifications
 */
-import * as lambda from "@aws-cdk/aws-lambda"; //Needed to avoid semgrep throwing up https://cwe.mitre.org/data/definitions/95.html
+import { LayerVersion, Runtime } from "@aws-cdk/aws-lambda";
 import { SnsEventSource } from "@aws-cdk/aws-lambda-event-sources";
+import { NodejsFunction } from "@aws-cdk/aws-lambda-nodejs";
 import { ITopic } from "@aws-cdk/aws-sns";
 import { Construct } from "@aws-cdk/core";
-import * as Path from "path";
+import { join } from "path";
 import { BuildConfig } from "../build/buildConfig";
 
 function name(buildConfig: BuildConfig, resourcename: string): string {
@@ -19,7 +20,7 @@ export interface SSOGroupProcessorProps {
   readonly groupsTableName: string;
   readonly linkManagerTopicArn: string;
   readonly errorNotificationsTopicArn: string;
-  readonly nodeJsLayer: lambda.LayerVersion;
+  readonly nodeJsLayer: LayerVersion;
   readonly ssoGroupEventNotificationsTopic: ITopic;
   readonly listInstancesSSOAPIRoleArn: string;
   readonly processTargetAccountSMInvokeRoleArn: string;
@@ -27,7 +28,7 @@ export interface SSOGroupProcessorProps {
 }
 
 export class SSOGroupProcessor extends Construct {
-  public readonly ssoGroupHandler: lambda.Function;
+  public readonly ssoGroupHandler: NodejsFunction;
 
   constructor(
     scope: Construct,
@@ -37,16 +38,31 @@ export class SSOGroupProcessor extends Construct {
   ) {
     super(scope, id);
 
-    this.ssoGroupHandler = new lambda.Function(
+    this.ssoGroupHandler = new NodejsFunction(
       this,
       name(buildConfig, "ssoGroupHandler"),
       {
-        runtime: lambda.Runtime.NODEJS_14_X,
-        handler: "groupsCud.handler",
+        runtime: Runtime.NODEJS_14_X,
         functionName: name(buildConfig, "ssoGroupHandler"),
-        code: lambda.Code.fromAsset(
-          Path.join(__dirname, "../", "lambda", "functions", "sso-handlers")
+        entry: join(
+          __dirname,
+          "../",
+          "lambda-functions",
+          "sso-handlers",
+          "src",
+          "groupsCud.ts"
         ),
+        bundling: {
+          externalModules: [
+            "@aws-sdk/client-dynamodb",
+            "@aws-sdk/lib-dynamodb",
+            "@aws-sdk/client-sns",
+            "@aws-sdk/client-sfn",
+            "@aws-sdk/client-sso-admin",
+            "@aws-sdk/credential-providers",
+          ],
+          minify: true,
+        },
         layers: [ssoGroupProcessorProps.nodeJsLayer],
         environment: {
           DdbTable: ssoGroupProcessorProps.groupsTableName,
